separating model from rendered creature
